Remove isLocal & renames
